Applied patch from Stefan Huehner (OFBIZ-514) to remove some unnecessary casts.
git-svn-id: https://svn.apache.org/repos/asf/incubator/ofbiz/trunk@482260 13f79535-47bb-0310-9956-ffa450edef68

+++ applications/order/src/org/ofbiz/order/order/OrderReadHelper.java
@@ -1690,7 +1690,7 @@ public class OrderReadHelper {
            for (Iterator groupiter = group.iterator(); groupiter.hasNext(); ) {
                GenericValue returnItem = (GenericValue) groupiter.next();
                if (returnItem.getDouble("returnQuantity") != null) {
-                   returned += ((Double) returnItem.getDouble("returnQuantity")).doubleValue();
+                   returned += (returnItem.getDouble("returnQuantity")).doubleValue();
                }
            }
